optimize

+++ tests/bootstrap.php
@@ -1,9 +1,17 @@
 <?php
 
-exec(PHP_BINARY.' '.__DIR__.'/setup.php', $output);
+$appPath = __DIR__.'/app.php';
 
-register_shutdown_function(function () {
-    $pidFile = __DIR__.'/../vendor/laravel/lumen/storage/lumen-swoole.pid';
+exec(__DIR__."/../bin/lumen-swoole -s $appPath -d");
+
+register_shutdown_function(function () use ($appPath) {
+    $app = require $appPath;
+    $pidFile = $app->storagePath('lumen-swoole.pid');
+
+    if (!file_exists($pidFile)) {
+        return;
+    }
+
     $pid = file_get_contents($pidFile);
 
     posix_kill($pid, SIGTERM);
